06/03/2026

+++ Vfi.Ui.Mvc.Vfi/Controllers/Authorization/UserController.cs
@@ -194,7 +194,7 @@ namespace Vfi.Ui.Mvc.Vfi.Controllers.Authorization {
             var listUserAccountModels = userAccounts.Select
             (user => new UserModel {
                 UserId = user.UserId,
-                Username = user.Username.ToString(),
+                Username = user.Username.ToString(),                      
                 Password = user.Password,
                 Active = user.Active != null && user.Active.Value,
                 Email = user.Email,
@@ -209,7 +209,8 @@ namespace Vfi.Ui.Mvc.Vfi.Controllers.Authorization {
             var model = new List<UserModel>();
             using (var vfi = new tammaContext()) {
                 model = (from x in vfi.Users
-                        where x.Active.Value
+                         //where x.Active.Value                           // 06/03/2026
+                        where x.Active == true                          //add 06/03/2026
                         orderby x.Username
                         select new UserModel { 
                             UserId =  x.UserId,
@@ -243,9 +244,11 @@ namespace Vfi.Ui.Mvc.Vfi.Controllers.Authorization {
                 // check if exist user account
                 if (result != null)
                     ModelState.AddModelError("Username", "Exist Username");
-                else if (ModelState.IsValid) {
+                //else if (ModelState.IsValid) {                    // 06/03/2026
+                else {                                                      // add 06/03/2026
                     //The model is valid - insert the user.
-                    var r = _userService.CreateNewUser(userModel);
+                    //var r = _userService.CreateNewUser(userModel);                // 06/03/2026
+                    _userService.CreateNewUser(userModel);                              // add 06/03/2026
                     _unitOfWork.SaveChanges();
                 }
             }
@@ -261,13 +264,15 @@ namespace Vfi.Ui.Mvc.Vfi.Controllers.Authorization {
         [GridAction]
         public ActionResult UpdateUser(int userId) {
             var user = _userService.GetUserById(userId);
-
+            if (user == null)
+                return HttpNotFound();
             var userModel = new UserModel { UserId = user.UserId };
 
             if (TryUpdateModel(userModel)) {
                 if (ModelState.IsValid) {
                     //The model is valid - edit the user.
-                    _userService.SaveUser(userModel);
+                    _userService.SaveUser(userModel);                 
+                    _unitOfWork.SaveChanges();                               //add 06/03/2026
                 }
             }
             else if (ModelState.IsValid == false) {
@@ -281,9 +286,11 @@ namespace Vfi.Ui.Mvc.Vfi.Controllers.Authorization {
         [GridAction]
         public ActionResult DeleteUser(int userId) {
             var user = _userService.GetUserById(userId);
-
+            if (user == null)
+                return HttpNotFound();
             try {
                 _userService.DeleteUser(userId);
+                _unitOfWork.SaveChanges();
             }
             catch (Exception) {
                 throw;
@@ -294,7 +301,6 @@ namespace Vfi.Ui.Mvc.Vfi.Controllers.Authorization {
         }
 
         public ActionResult SelectComboBoxSalesUser() {
-
             using (var vfi = new tammaContext()) {
                 return new JsonResult {
                     Data = new SelectList(
